Move special gear into own subdirectory
